# docs/swagger.yaml
definitions:
  domain.Credentials:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Авторизует пользователя в системе и возвращает токен авторизации
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/domain.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация, возвращает статус и токен в заголовке
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка в формате запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Неверный логин или пароль
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация пользователя
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя в системе и возвращает токен авторизации
      parameters:
      - description: Учетные данные пользователя
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/domain.Credentials'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная регистрация, возвращает статус и токен в заголовке
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка в формате запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Пользователь с таким логином уже существует
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация нового пользователя
      tags:
      - auth
swagger: "2.0"